Redirections - Search and frontend

+++ routes/admin/admin.php
@@ -9,5 +9,7 @@ Route::get('/admin', [AdminController::class, 'index'])->name('admin');
 Route::get('/admin/redirections', [RedirectionController::class, 'index'])->name('admin.redirections');
 Route::get('/admin/redirections/create', [RedirectionController::class, 'create'])->name('admin.redirections.create');
 Route::post('/admin/redirections/create', [RedirectionController::class, 'store']);
-Route::delete('/admin/redirections/{redirection}/delete', [RedirectionController::class, 'destroy']);
+
 Route::get('/admin/redirections/{redirection}/edit', [RedirectionController::class, 'edit'])->name('admin.redirections.edit');
+Route::patch('/admin/redirections/{redirection}/edit', [RedirectionController::class, 'update']);
+Route::delete('/admin/redirections/{redirection}/delete', [RedirectionController::class, 'destroy'])->name('admin.redirections.delete');
